refactor: declare type decoration of fastify instance in plugin files
