+++ @types/fastify/fastify.d.ts
@@ -1,0 +1,7 @@
+import { IAuth } from "../../src/schemas/auth.ts";
+
+declare module "fastify" {
+    export interface FastifyRequest {
+      user: IAuth
+    }
+  }
