+++ src/routes/api/autohooks.ts
@@ -1,5 +1,6 @@
 import { FastifyInstance } from "fastify";
 
+
 export default async function (fastify: FastifyInstance) {
   fastify.addHook("onRequest", async (request) => {
     if (!request.url.startsWith("/api/auth/login")) {
